Re-add partial write tests
